+++ src/components/CheckoutForm/CheckoutForm.module.scss
@@ -22,24 +22,6 @@
 
 .confirmation {
   grid-area: confirmation;
-
-  &__form {
-    height: calc(100% - 36px);
-  }
-
-  &__dai {
-    @include text('caption');
-  }
-
-  &__contact {
-    margin-left: 4px;
-    text-decoration: underline;
-    color: var(--jungle-green);
-
-    &:hover {
-      color: var(--jungle-green-hover);
-    }
-  }
 }
 
 .top {
@@ -92,28 +74,6 @@
 
 .row {
   display: flex;
-}
-
-.input {
-  height: 40px;
-
-  &:hover,
-  &:focus {
-    border-color: var(--jungle-green);
-  }
-
-  &_right {
-    margin-left: 10px;
-  }
-
-  &_address {
-    margin: 0 10px;
-  }
-}
-
-.discount {
-  position: relative;
-  margin-bottom: 16px;
 }
 
 .label_card {
@@ -222,110 +182,16 @@
   text-align: center;
 }
 
-.plan {
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 0 0 16px;
-
-  &__check {
-    box-sizing: content-box;
-    fill: var(--jungle-green);
-    padding: 12px 11px;
-    height: 7.5px;
-    width: 10px;
-    background: #ccf2e7;
-    border-radius: 50%;
-    margin-right: 12px;
-  }
-
-  &__left {
-    @include text('body-1', 'm');
-
-    color: var(--mirage);
-    display: flex;
-    align-items: center;
-  }
-
-  &__right {
-    color: var(--casper);
-    text-align: right;
-
-    @include text('caption');
-  }
-
-  &__year {
-    @include text('h4', 'm');
-
-    color: var(--waterloo);
-  }
-
-  &__month {
-    @include text('body-3', 'm');
-
-    color: var(--waterloo);
-  }
-}
-
-.hold {
-  position: relative;
-  border-radius: 4px;
-  color: var(--mirage);
-  padding: 12px 16px 8px 40px;
-  background: #ecfaf6;
-  margin-bottom: 44px;
-
-  &__icon {
-    position: absolute;
-    left: 17px;
-    top: 15px;
-  }
-}
-
-.btn {
+.input {
   height: 40px;
-  justify-content: center;
-  margin: 37px 0 16px;
-}
+  margin-bottom: 16px;
 
-.expl {
-  @include text('caption');
-
-  color: var(--waterloo);
-  text-align: center;
-}
-
-.check {
-  color: var(--waterloo);
-
-  &__label {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 12px;
+  &:hover,
+  &:focus {
+    border-color: var(--jungle-green);
   }
 
-  &__discount {
-    color: var(--jungle-green);
+  &_last {
+    margin: 0;
   }
-
-  &__total {
-    display: flex;
-    justify-content: space-between;
-    color: var(--mirage);
-    padding-top: 25px;
-    border-top: 1px solid var(--porcelain);
-    margin-top: 24px;
-  }
-
-  &__price {
-    @include text('body-1', 'm');
-  }
-}
-
-.valid {
-  position: absolute;
-  top: 50%;
-  right: 16px;
-  transform: translateY(-50%);
 }
